Split implementation of GoalHealthKitConnection
This separates out code which is only relevant to quantity or category
metrics into the relevant classes, with shared code in the base class.
For now this uses a classic "inheritance" pattern with the base class
calling into some methods of the subclasses. This should be cleaned up
into a mix of delegate/helper, but as a first pass this gets the code
separated.

This largely gets rid of the optional metric types in all this area of code

Test Plan:
Compiled the code
